+++ Main.py
@@ -1,53 +1,14 @@
+from Client import get_ai_client
+
 from openai import OpenAI
 import re
-import os
-from dotenv import load_dotenv
 
-system_prompt = open("SystemMessage.txt", "r").read()
+food_macros = open("GptSystemMessages/FoodMacros.txt", "r").read()
 
-def is_client_valid(c: OpenAI):
-    try:
-        c.chat.completions.create(
-            model="gpt-3.5-turbo",
-            messages=[{"role": "user", "content": "Say this is a test"}],
-            max_tokens=1
-        )
-    except:
-        return False
-    else:
-        return True
-    
-
-def get_ai_client() -> OpenAI:
-    load_dotenv()
-    if os.environ.get("OPENAI_API_KEY"):
-        c = OpenAI(
-            api_key=os.environ.get("OPENAI_API_KEY")
-        )
-        if is_client_valid(c):
-            print("OpenAI API key loaded from environment.")
-            return c
-        
-    user_key = input("Input OpenAI API key: ")
-    if user_key == "":
-        return
-    
-    c = OpenAI(
-        api_key=user_key
-    )
-    valid = is_client_valid(c)
-    if valid:
-        print("AI successfully loaded!")
-        return c
-    else:
-        print("Invalid key, try again...")
-        return get_ai_client()
-    
-
-def prompt_ai(client: OpenAI, food: str):
+def prompt_ai_for_macros(client: OpenAI, food: str):
     return client.chat.completions.create(
         model="gpt-3.5-turbo",
-        messages=[{"role": "system", "content": system_prompt}, {"role": "user", "content": food}],
+        messages=[{"role": "system", "content": food_macros}, {"role": "user", "content": food}],
     )
 
 
@@ -67,7 +28,7 @@ def get_macros_prompt(client: OpenAI, user_food: str):
     attempts = 3
     while attempts > 0:
         attempts -= 1
-        result = prompt_ai(client, user_food)
+        result = prompt_ai_for_macros(client, user_food)
         macros = get_protein_and_calories(result.choices[0].message.content)
         if macros:
             print("Total Protein: " + str(macros[0]) + "g")
@@ -84,9 +45,16 @@ def main():
     
     if client:
         while True:
-            user_input = input("Describe the food: ")
-            if user_input:
-                get_macros_prompt(client, user_input)
+            mode_input = input("Selected mode (image = 'i', text = 't'): ")
+            if mode_input == "i":
+                print("temp")
+            elif mode_input == "t":
+                while True:
+                    user_input = input("Describe the food: ")
+                    if user_input:
+                        get_macros_prompt(client, user_input)
+                    else:
+                        break
             else:
                 break
